// src/types/task.ts
export interface Task {
  id: string;
  mode: 'Planifié automatiquement' | 'Manuel';
  coordination: string;
  marketType: string;
  method: string;
  conduct: string;
  reference: string;
  description: string;
  taskName: string;
  resources: string;
  status: 'Non commencé' | 'En cours' | 'Terminé' | 'Annulé';
  physicalProgress: number;
  plannedDuration: number;
  plannedStart: string | null;
  plannedEnd: string | null;
  completionPercentage: number;
  startAlert: boolean;
  startAlertSent: boolean;
  startAlertDate: string | null;
  actualDuration: number;
  actualStart: string | null;
  actualEnd: string | null;
  actualProgress: number;
  remainingDays: number;
  actualStartAlert: boolean;
  actualStartAlertSent: boolean;
  actualStartAlertDate: string | null;
  endAlert: boolean;
  endAlertSent: boolean;
  endAlertDate: string | null;
  startGap: number;
  durationGap: number;
  comment: string;
  commentDate: string | null;
  reminder: boolean;
  reminderDate: string | null;
  plannedReminder: string | null;
  cancelled: boolean;
  requiredDocuments: string[];
  predecessors: string[];
}